new mouse cursor for intersection operations.
	* cursors/{mouse1_u,mouse1_umsk}: new mouse cursor for intersection
	operations.

	* app/cursorutil.[ch], app/rect_select.c: use the new cursor.

	* app/gimpimage.c:  Applied layer removal bug fix from
	David Le Corfec, <lecorfec@etudiant.univ-mlv.fr>

	* plug-ins/gdyntext/{font_selection.c, gdyntext.c, gdyntext_ui.c}:
	replaced snprintf with g_snprintf.

	* plug-ins/jpeg/jpeg.c: updated to work with the double precision
	resolutions.

+++ plug-ins/common/jpeg.c
@@ -855,8 +855,8 @@ save_image (char   *filename,
 
 #ifdef GIMP_HAVE_RESOLUTION_INFO
   {
-    float xresolution;
-    float yresolution;
+    double xresolution;
+    double yresolution;
 
     gimp_image_get_resolution (image_ID, &xresolution, &yresolution);
